Removing fastapi imports from standard import tree
(cherry picked from commit 4c96cb8647096395395e167033da69d2186d6366)
