+++ json_logging/__init__.py
@@ -400,10 +400,11 @@ def init_connexion(custom_formatter=None, enable_json=False):
 # register FastAPI support
 import json_logging.framework.fastapi as fastapi_support
 
-register_framework_support('fastapi', app_configurator=None,
-                           app_request_instrumentation_configurator=fastapi_support.FastAPIAppRequestInstrumentationConfigurator,
-                           request_adapter_class=fastapi_support.FastAPIRequestAdapter,
-                           response_adapter_class=fastapi_support.FastAPIResponseAdapter)
+if fastapi_support.is_fastapi_present():
+    register_framework_support('fastapi', app_configurator=None,
+                               app_request_instrumentation_configurator=fastapi_support.FastAPIAppRequestInstrumentationConfigurator,
+                               request_adapter_class=fastapi_support.FastAPIRequestAdapter,
+                               response_adapter_class=fastapi_support.FastAPIResponseAdapter)
 
 def init_fastapi(custom_formatter=None, enable_json=False):
     __init(framework_name='fastapi', custom_formatter=custom_formatter, enable_json=enable_json)
